rename  baaImageLoadFunction.provider > olLoadFunction.provider
